nspawn: support custom container service name
We were hardcoding "systemd-nspawn" as the value of the $container env
variable and "nspawn" as the service string in machined registration.

This commit allows the user to configure it by setting the $container
env variable when calling systemd-nspawn.

If $container is not set, we use the string "systemd-nspawn" for both,
fixing the previous inconsistency.